+++ apps/ledger-live-mobile/src/components/GenericErrorView.tsx
@@ -9,14 +9,14 @@ import TranslatedError from "./TranslatedError";
 import SupportLinkError from "./SupportLinkError";
 import BluetoothDisabled from "./RequiresBLE/BluetoothDisabled";
 import { GenericInformationBody } from "./GenericInformationBody";
-import { isDmkError, isiOSPeerRemovedPairingError } from "@ledgerhq/live-dmk-mobile";
+import { DmkError, isDmkError, isiOSPeerRemovedPairingError } from "@ledgerhq/live-dmk-mobile";
 
 type Props = {
-  error: Error;
+  error: Error | DmkError;
   // sometimes we want to "hide" the technical error into a category
   // for instance, for Genuine check we want to express "Genuine check failed" because "<actual error>"
   // in such case, the outerError is GenuineCheckFailed and the actual error is still error
-  outerError?: Error | null;
+  outerError?: Error | DmkError | null;
   withDescription?: boolean;
   withHelp?: boolean;
   hasExportLogButton?: boolean;