+++ apps/ledger-live-mobile/src/const/navigation.ts
@@ -496,6 +496,30 @@ export enum ScreenName {
   NearWithdrawingConnectDevice = "NearWithdrawingConnectDevice",
   NearWithdrawingValidationError = "NearWithdrawingValidationError",
   NearWithdrawingValidationSuccess = "NearWithdrawingValidationSuccess",
+  // Aptos
+  AptosStakingValidator = "AptosStakingValidator",
+  AptosStakingValidatorSelect = "AptosStakingValidatorSelect",
+  AptosStakingAmount = "AptosStakingAmount",
+  AptosStakingStarted = "AptosStakingStarted",
+  AptosStakingSelectDevice = "AptosStakingSelectDevice",
+  AptosStakingConnectDevice = "AptosStakingConnectDevice",
+  AptosStakingValidationError = "AptosStakingValidationError",
+  AptosStakingValidationSuccess = "AptosStakingValidationSuccess",
+  AptosRestakingAmount = "AptosRestakingAmount",
+  AptosRestakingSelectDevice = "AptosRestakingSelectDevice",
+  AptosRestakingConnectDevice = "AptosRestakingConnectDevice",
+  AptosRestakingValidationError = "AptosRestakingValidationError",
+  AptosRestakingValidationSuccess = "AptosRestakingValidationSuccess",
+  AptosUnstakingAmount = "AptosUnstakingAmount",
+  AptosUnstakingSelectDevice = "AptosUnstakingSelectDevice",
+  AptosUnstakingConnectDevice = "AptosUnstakingConnectDevice",
+  AptosUnstakingValidationError = "AptosUnstakingValidationError",
+  AptosUnstakingValidationSuccess = "AptosUnstakingValidationSuccess",
+  AptosWithdrawingAmount = "AptosWithdrawingAmount",
+  AptosWithdrawingSelectDevice = "AptosWithdrawingSelectDevice",
+  AptosWithdrawingConnectDevice = "AptosWithdrawingConnectDevice",
+  AptosWithdrawingValidationError = "AptosWithdrawingValidationError",
+  AptosWithdrawingValidationSuccess = "AptosWithdrawingValidationSuccess",
 
   Newsfeed = "Newsfeed",
 
@@ -533,6 +557,7 @@ export enum ScreenName {
   LedgerSyncDeepLinkHandler = "LedgerSyncDeepLinkHandler",
 
   MockedAddAssetButton = "MockedAddAssetButton",
+  MockedModularDrawer = "MockedModularDrawer",
   MockedWalletScreen = "MockedWalletScreen",
   GenericLandingPage = "GenericLandingPage",
 
@@ -644,6 +669,11 @@ export enum NavigatorName {
   NearStakingFlow = "NearStakingFlow",
   NearUnstakingFlow = "NearUnstakingFlow",
   NearWithdrawingFlow = "NearWithdrawingFlow",
+  // Aptos
+  AptosStakingFlow = "AptosStakingFlow",
+  AptosUnstakingFlow = "AptosUnstakingFlow",
+  AptosWithdrawingFlow = "AptosWithdrawingFlow",
+  AptosRestakingFlow = "AptosRestakingFlow",
 
   // Tab
   Main = "Main",