+++ apps/ledger-live-mobile/src/newArch/features/ModularDrawer/ModularDrawer.tsx
@@ -1,33 +1,99 @@
-import React from "react";
+import React, { useState } from "react";
 import QueuedDrawer from "~/components/QueuedDrawer";
 import ModularDrawerFlowManager from "./ModularDrawerFlowManager";
 import { ModularDrawerStep } from "./types";
 import { useModularDrawerFlowStepManager } from "./hooks/useModularDrawerFlowStepManager";
 
-type Props = {
-  selectedStep: ModularDrawerStep;
-  isOpen?: boolean;
-  onClose?: () => void;
-};
-export function ModularDrawer({ isOpen, onClose, selectedStep }: Props) {
-  const viewModel = useModularDrawerFlowStepManager({ selectedStep });
+import { useInitModularDrawer } from "./hooks/useInitModularDrawer";
+import { useAssets } from "./hooks/useAssets";
+import { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
+import { useModularDrawerState } from "./hooks/useModularDrawerState";
 
-  // In prevision of tracking the back button press
-  const onBackButtonPress = () => viewModel.prevStep();
-  const onCloseButtonPress = () => {
+/**
+ * Props for the ModularDrawer component.
+ */
+type ModularDrawerProps = {
+  /**
+   * The current step to display in the drawer navigation flow.
+   */
+  selectedStep: ModularDrawerStep;
+  /**
+   * Whether the drawer is open.
+   */
+  isOpen?: boolean;
+  /**
+   * Callback fired when the drawer is closed.
+   */
+  onClose?: () => void;
+  /**
+   * List of currencies to display in the drawer.
+   */
+  currencies: CryptoOrTokenCurrency[];
+};
+/**
+ * ModularDrawer is a generic drawer component for asset/network selection flows.
+ * Handles navigation steps, asset/network selection, and drawer state.
+ *
+ * @param {ModularDrawerProps} props - The props for the ModularDrawer component.
+ */
+export function ModularDrawer({ isOpen, onClose, selectedStep, currencies }: ModularDrawerProps) {
+  const navigationStepManager = useModularDrawerFlowStepManager({ selectedStep });
+  const [defaultSearchValue, setDefaultSearchValue] = useState("");
+  const [itemsToDisplay, setItemsToDisplay] = useState<CryptoOrTokenCurrency[]>([]);
+
+  const { sortedCryptoCurrencies, isReadyToBeDisplayed, currenciesByProvider } =
+    useInitModularDrawer();
+
+  const { availableAssets, currencyIdsArray } = useAssets(currencies, sortedCryptoCurrencies);
+
+  const { handleAsset, handleNetwork, reset, handleBack, availableNetworks } =
+    useModularDrawerState({
+      goToStep: navigationStepManager.goToStep,
+      currenciesByProvider,
+      currencyIds: currencyIdsArray,
+    });
+
+  /**
+   * Handlers for the back & close button in the drawer.
+   */
+
+  const handleBackButton = () => {
+    handleBack(navigationStepManager.currentStep);
+  };
+
+  const handleCloseButton = () => {
     onClose?.();
-    viewModel.reset();
+    navigationStepManager.reset();
+    reset();
   };
 
   return (
     <QueuedDrawer
       isRequestingToBeOpened={isOpen}
-      onClose={onCloseButtonPress}
-      hasBackButton={viewModel.hasBackButton}
-      onBack={onBackButtonPress}
+      onClose={handleCloseButton}
+      hasBackButton={navigationStepManager.hasBackButton}
+      onBack={handleBackButton}
+      containerStyle={{
+        maxHeight: "90%",
+      }}
     >
       {/* TODO: Drawer Transitions & animations will be implemented here. */}
-      <ModularDrawerFlowManager {...viewModel} />
+      <ModularDrawerFlowManager
+        navigationStepViewModel={navigationStepManager}
+        assetsViewModel={{
+          availableAssets,
+          onAssetSelected: handleAsset,
+          defaultSearchValue,
+          setDefaultSearchValue,
+          itemsToDisplay,
+          setItemsToDisplay,
+        }}
+        networksViewModel={{
+          onNetworkSelected: handleNetwork,
+          availableNetworks,
+        }}
+        isReadyToBeDisplayed={isReadyToBeDisplayed}
+      />
     </QueuedDrawer>
   );
 }