+++ e2e/mobile/specs/portfolio/portfolioAccountsTab.spec.ts
@@ -1,0 +1,19 @@
+const testConfig = {
+  tmsLinks: ["B2CQA-2871"],
+  tags: ["@NanoSP", "@LNS", "@NanoX"],
+};
+
+describe("Wallet Page", () => {
+  beforeAll(async () => {
+    await app.init({
+      userdata: "speculos-tests-app",
+    });
+    await app.portfolio.waitForPortfolioPageToLoad();
+  });
+
+  testConfig.tmsLinks.forEach(link => $TmsLink(link));
+  testConfig.tags.forEach(tag => $Tag(tag));
+  it("Portfolio Accounts Tab - LLM", async () => {
+    await app.portfolio.checkAccountsSection();
+  });
+});