+++ e2e/mobile/specs/swap/otherTestCases/swap.other.ts
@@ -222,10 +222,13 @@ export function runTooLowAmountForQuoteSwapsTest(
         swap.accountToDebit,
         swap.accountToCredit,
       );
+
+      const actualAmount = swap.amount === "USE_MIN_AMOUNT" ? minAmount : swap.amount;
+
       await performSwapUntilQuoteSelectionStep(
         swap.accountToDebit,
         swap.accountToCredit,
-        minAmount,
+        actualAmount,
         quotesVisible,
       );
       if (quotesVisible) {
@@ -282,6 +285,7 @@ export function runUserRefusesTransactionTest(
 
       await checkSwapInfosOnDeviceVerificationStep(rejectedSwap, selectedProvider, minAmount);
       await app.swap.verifyAmountsAndRejectSwap(rejectedSwap, minAmount);
+      await app.swap.verifyDeviceActionLoadingNotVisible();
       await app.swapLiveApp.checkErrorMessage("User refused");
     });
   });