+++ libs/coin-modules/coin-sui/src/network/sdk.ts
@@ -1,9 +1,10 @@
 import {
-  PaginatedTransactionResponse,
-  SuiClient,
   ExecuteTransactionBlockParams,
-  TransactionEffects,
+  PaginatedTransactionResponse,
   QueryTransactionBlocksParams,
+  SuiClient,
+  SuiHTTPTransport,
+  TransactionEffects,
 } from "@mysten/sui/client";
 import { TransactionBlockData, SuiTransactionBlockResponse, SuiCallArg } from "@mysten/sui/client";
 import { Transaction } from "@mysten/sui/transactions";
@@ -19,6 +20,7 @@ import type { CreateExtrinsicArg } from "../logic/craftTransaction";
 import { ensureAddressFormat } from "../utils";
 import coinConfig from "../config";
 import { SuiAsset } from "../api/types";
+import { getEnv } from "@ledgerhq/live-env";
 
 type AsyncApiFunction<T> = (api: SuiClient) => Promise<T>;
 
@@ -30,15 +32,32 @@ const BLOCK_HEIGHT = 5; // sui has no block height metainfo, we use it simulate 
 
 export const DEFAULT_COIN_TYPE = "0x2::sui::SUI";
 
+type GenericInput<T> = T extends (...args: infer K) => unknown ? K : never;
+type Inputs = GenericInput<typeof fetch>;
+
+const fetcher = (url: Inputs[0], options: Inputs[1], retry = 3): Promise<Response> => {
+  if (options) {
+    options.headers = {
+      ...options.headers,
+      "X-Ledger-Client-Version": getEnv("LEDGER_CLIENT_VERSION"),
+    };
+  }
+  if (retry === 1) return fetch(url, options);
+
+  return fetch(url, options).catch(() => fetcher(url, options, retry - 1));
+};
+
 /**
  * Connects to Sui Api
  */
 export async function withApi<T>(execute: AsyncApiFunction<T>) {
   const url = coinConfig.getCoinConfig().node.url;
+  const transport = new SuiHTTPTransport({
+    url,
+    fetch: fetcher,
+  });
 
-  if (!apiMap[url]) {
-    apiMap[url] = new SuiClient({ url });
-  }
+  apiMap[url] ??= new SuiClient({ transport });
 
   const result = await execute(apiMap[url]);
   return result;
@@ -229,6 +248,8 @@ function transactionToOp(address: string, transaction: SuiTransactionBlockRespon
       block: {
         // agreed to return bigint
         height: BigInt(transaction.checkpoint || "") as unknown as number,
+        hash,
+        time: getOperationDate(transaction),
       },
     },
     asset: { type: "native" },
@@ -253,27 +274,83 @@ export const getLastBlock = () =>
 export const getOperations = async (
   accountId: string,
   addr: string,
-  cursor?: string | null | undefined,
+  cursor?: QueryTransactionBlocksParams["cursor"],
 ): Promise<Operation[]> =>
   withApi(async api => {
-    const sentOps = await loadOperations({ api, addr, type: "OUT", cursor });
-    const receivedOps = await loadOperations({ api, addr, type: "IN", cursor });
-    const rawTransactions = [...sentOps, ...receivedOps].sort(
-      (a, b) => Number(b.timestampMs) - Number(a.timestampMs),
-    );
+    const sentOps = await loadOperations({
+      api,
+      addr,
+      type: "OUT",
+      cursor,
+      order: cursor ? "ascending" : "descending",
+      operations: [],
+    });
+    const receivedOps = await loadOperations({
+      api,
+      addr,
+      type: "IN",
+      cursor,
+      order: cursor ? "ascending" : "descending",
+      operations: [],
+    });
+    const rawTransactions = filterOperations(sentOps, receivedOps, cursor);
 
     return rawTransactions.map(transaction => transactionToOperation(accountId, addr, transaction));
   });
 
-export const getListOperations = async (addr: string, cursor = ""): Promise<Op<SuiAsset>[]> =>
-  withApi(async api => {
-    const opsOut = await loadOperations({ api, addr, type: "OUT", cursor });
-    const opsIn = await loadOperations({ api, addr, type: "IN", cursor });
+export const filterOperations = (
+  operationList1: SuiTransactionBlockResponse[],
+  operationList2: SuiTransactionBlockResponse[],
+  cursor: string | null | undefined,
+) => {
+  let filterTimestamp = 0;
 
-    const rawTransactions = [...opsIn, ...opsOut].sort(
-      (a, b) => Number(b.timestampMs) - Number(a.timestampMs),
-    );
-    const list = uniqBy(rawTransactions, tx => tx.digest);
+  // When restoring state (no cursor provided) and we've reached the limit for either sent or received operations,
+  // we filter out extra operations to maintain correct chronological order
+  if (
+    !cursor &&
+    operationList1.length &&
+    operationList2.length &&
+    (operationList1.length === TRANSACTIONS_LIMIT || operationList2.length === TRANSACTIONS_LIMIT)
+  ) {
+    const aTime = operationList1[operationList1.length - 1].timestampMs ?? 0;
+    const bTime = operationList2[operationList2.length - 1].timestampMs ?? 0;
+    filterTimestamp = Math.max(Number(aTime), Number(bTime));
+  }
+  const result = [...operationList1, ...operationList2]
+    .sort((a, b) => Number(b.timestampMs) - Number(a.timestampMs))
+    .filter(op => Number(op.timestampMs) >= filterTimestamp);
+
+  return uniqBy(result, tx => tx.digest);
+};
+
+/**
+ * Fetch operations for Alpaca
+ */
+export const getListOperations = async (
+  addr: string,
+  cursor: QueryTransactionBlocksParams["cursor"] = null,
+  withApiImpl: typeof withApi = withApi,
+): Promise<Op<SuiAsset>[]> =>
+  withApiImpl(async api => {
+    const opsOut = await loadOperations({
+      api,
+      addr,
+      type: "OUT",
+      cursor,
+      order: cursor ? "ascending" : "descending",
+      operations: [],
+    });
+    const opsIn = await loadOperations({
+      api,
+      addr,
+      type: "IN",
+      cursor,
+      order: cursor ? "ascending" : "descending",
+      operations: [],
+    });
+    const list = filterOperations(opsIn, opsOut, cursor);
+
     return list.map(t => transactionToOp(addr, t));
   });
 
@@ -373,45 +450,44 @@ export const executeTransactionBlock = async (params: ExecuteTransactionBlockPar
 /**
  * Fetch operations for a specific address and type until the limit is reached
  */
-export const loadOperations = async (params: {
+export const loadOperations = async ({
+  cursor,
+  operations,
+  order,
+  ...params
+}: {
   api: SuiClient;
   addr: string;
   type: OperationType;
-  cursor?: string | null | undefined;
+  operations: PaginatedTransactionResponse["data"];
+  order: "ascending" | "descending";
+  cursor?: QueryTransactionBlocksParams["cursor"];
 }): Promise<PaginatedTransactionResponse["data"]> => {
-  const operations: PaginatedTransactionResponse["data"] = [];
-  let currentCursor = params.cursor;
+  try {
+    if (order === "descending" && operations.length >= TRANSACTIONS_LIMIT) {
+      return operations;
+    }
 
-  while (operations.length < TRANSACTIONS_LIMIT) {
-    try {
-      const { data, nextCursor, hasNextPage } = await queryTransactions({
-        ...params,
-        cursor: currentCursor,
+    const { data, nextCursor, hasNextPage } = await queryTransactions({
+      ...params,
+      order,
+      cursor,
+    });
+
+    operations.push(...data);
+    if (!hasNextPage) {
+      return operations;
+    }
+
+    await loadOperations({ ...params, cursor: nextCursor, operations, order });
+  } catch (error: any) {
+    if (error.type === "InvalidParams") {
+      log("coin:sui", "(network/sdk): loadOperations failed with cursor, retrying without it", {
+        error,
+        params,
       });
-
-      operations.push(...data);
-
-      // If we got fewer results than the query limit or no more data, we've reached the end
-      if (data.length < TRANSACTIONS_LIMIT_PER_QUERY || !hasNextPage) {
-        break;
-      }
-
-      currentCursor = nextCursor;
-    } catch (error: any) {
-      if (error.type === "InvalidParams") {
-        log("coin:sui", "(network/sdk): loadOperations failed with cursor, retrying without it", {
-          error,
-          params,
-        });
-
-        currentCursor = null;
-
-        continue;
-      } else {
-        log("coin:sui", "(network/sdk): loadOperations error", { error, params });
-
-        break;
-      }
+    } else {
+      log("coin:sui", "(network/sdk): loadOperations error", { error, params });
     }
   }
 
@@ -425,16 +501,17 @@ export const queryTransactions = async (params: {
   api: SuiClient;
   addr: string;
   type: OperationType;
-  cursor?: string | null | undefined;
+  order: "ascending" | "descending";
+  cursor?: QueryTransactionBlocksParams["cursor"];
 }): Promise<PaginatedTransactionResponse> => {
-  const { api, addr, type, cursor } = params;
+  const { api, addr, type, cursor, order } = params;
   const filter: QueryTransactionBlocksParams["filter"] =
     type === "IN" ? { ToAddress: addr } : { FromAddress: addr };
 
   return await api.queryTransactionBlocks({
     filter,
     cursor,
-    order: "descending",
+    order,
     options: {
       showInput: true,
       showBalanceChanges: true,