+++ libs/ui/packages/react/src/pre-ldls/components/AccountItem/AccountItem.tsx
@@ -1,10 +1,11 @@
 import React from "react";
 import styled from "styled-components";
-import type { CheckboxProps } from "../../../components/form/Checkbox/Checkbox";
-import { Checkbox, Icon, Text, Flex } from "../../../components";
+import { Icon, Text, Flex } from "../../../components";
 import { withTokens } from "../../libs";
 import { Address } from "../Address/Address";
 import { Tag } from "../Tag/Tag";
+import type { CheckboxProps } from "../Checkbox/Checkbox";
+import { Checkbox } from "../Checkbox/Checkbox";
 
 export type Account = {
   address: string;