+++ libs/ledgerjs/packages/hw-transport-mocker/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/hw-transport-mocker
 
+## 6.29.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.8
+
+## 6.29.8-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.8-next.0
+
 ## 6.29.7
 
 ### Patch Changes